Time: 42 ms (56.61%), Space: 59.4 MB (50.92%) - LeetHub

+++ 2665-minimum-time-to-repair-cars/2665-minimum-time-to-repair-cars.cpp
@@ -8,7 +8,7 @@ public:
             long long mid = start + (end - start) / 2;
             long long totalCars = 0;
             for(int rank : ranks) {
-                totalCars += (long long)(sqrt((double)mid / rank));
+                totalCars += (long long)(sqrt(mid / rank));
             }
 
             if(totalCars >= cars) {
